Merge develop into master 2019/02/01

+++ app/Http/Controllers/ActionsController.php
@@ -27,6 +27,8 @@ class ActionsController extends Controller
 
     public function create(Objective $objective)
     {
+        $this->authorize('storeObjective', $objective->model);
+
         $priorities = Priority::all();
         $user = User::where('id', '=', auth()->user()->id)->first();
         $keyresults = KeyResult::where('objective_id', '=', $objective->id)->get();
@@ -42,6 +44,8 @@ class ActionsController extends Controller
 
     public function store(ActionRequest $request)
     {
+        $this->authorize('storeObjective', KeyResult::find($request->krs_id)->objective->model);
+
         $attr['user_id'] = auth()->user()->id;
         $attr['related_kr'] = $request->input('krs_id');
         $attr['priority'] = $request->input('priority');
@@ -77,6 +81,8 @@ class ActionsController extends Controller
 
     public function edit(Action $action)
     {
+        $this->authorize('update', $action);
+
         $priorities = Priority::all();
         $user = User::where('id', '=', auth()->user()->id)->first();
 
@@ -101,8 +107,9 @@ class ActionsController extends Controller
 
     public function update(ActionRequest $request, Action $action)
     {
-        
-        if ($request->input('invite') && $request->input('invite') != $action->user_id){
+        $this->authorize('update', $action);
+
+        if ($request->input('invite') && $request->input('invite') != $action->user_id) {
             $action->sendInvitation($request);
         }
 
@@ -125,18 +132,27 @@ class ActionsController extends Controller
 
     public function destroy(Action $action)
     {
+        $this->authorize('delete', $action);
+        $objective = $action->objective;
+        $redirectURL = $objective->model->getOKrRoute();
+        $action->invitation()->delete();
         $action->delete();
-        return redirect()->back();
+
+        return redirect()->to($redirectURL . '#oid-' . $objective->id);
     }
 
     public function destroyFile(Action $action, Media $media)
     {
+        $this->authorize('delete', $action);
+
         $media->delete();
         return redirect()->route('actions.edit', $action);
     }
 
     public function done(Action $action)
     {
+        $this->authorize('update', $action);
+
         $act = Action::find($action->id);
         $act->isdone = 'true';
         $act->save();
@@ -149,7 +165,6 @@ class ActionsController extends Controller
         return response()->json($results);
     }
 
-
     /**
      * 拒絕邀請
      *
@@ -160,7 +175,7 @@ class ActionsController extends Controller
     public function rejectInvite(Action $action, User $member)
     {
         $action->deleteInvitation($member);
-        return redirect()->route('user.action',$member->id);
+        return redirect()->route('user.action', $member->id);
     }
 
     /**
@@ -175,7 +190,7 @@ class ActionsController extends Controller
         $action->deleteInvitation($member);
         $attr['user_id'] = $member->id;
         $action->update($attr);
-        return redirect()->route('user.action',$member->id);
+        return redirect()->route('user.action', $member->id);
     }
