+++ app/Http/Controllers/ProjectController.php
@@ -6,8 +6,8 @@ use Illuminate\Http\Request;
 use App\Project;
 use App\Http\Requests\ObjectiveRequest;
 use App\User;
-use App\ProjectUser;
 use App\Invitation;
+use App\Permission;
 
 class ProjectController extends Controller
 {
@@ -17,6 +17,7 @@ class ProjectController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
+        $this->authorizeResource(Project::class, 'project');
     }
 
     /**
@@ -26,12 +27,12 @@ class ProjectController extends Controller
      */
     public function index(Request $request)
     {
-        $projects = Project::where('isdone', false)->get();
+        $projects = auth()->user()->projects->where('isdone', false);
         foreach ($projects as $project) {
             $project['okrs'] = $project->getOkrsWithPage($request)['okrs'];
         }
 
-        $projectDone = Project::where('isdone', true)->get();
+        $projectDone = auth()->user()->projects->where('isdone', true);
         foreach ($projectDone as $project) {
             $project['okrs'] = $project->getOkrsWithPage($request)['okrs'];
         }
@@ -46,16 +47,6 @@ class ProjectController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        return view('project.create');
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -65,35 +56,14 @@ class ProjectController extends Controller
     {
         $attr['name'] = $request->project_name;
         $attr['description'] = $request->project_description;
-        $attr['user_id'] = auth()->user()->id;
+        $attr['company_id'] = auth()->user()->company_id;
 
         $project = Project::create($attr);
         $project->addAvatar($request);
-        ProjectUser::create(['project_id' => $project->id, 'user_id' => auth()->user()->id]);
+        $project->users()->attach(auth()->user());
+        $project->createPermission(1);
 
         return redirect()->route('project');
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Project $project
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Project $project)
-    {
-        return view('project.edit', ['project' => $project]);
     }
 
     /**
@@ -111,7 +81,7 @@ class ProjectController extends Controller
 
         $project->addAvatar($request);
 
-        return redirect()->route('project');
+        return redirect()->back();
     }
 
     /**
@@ -122,7 +92,8 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        ProjectUser::where('project_id', $project->id)->delete();
+        $project->users()->detach();
+        $project->invitation()->delete();
         $project->delete();
 
         return redirect('project');
@@ -135,12 +106,13 @@ class ProjectController extends Controller
      */
     public function listOKR(Request $request, Project $project)
     {
+        $this->authorize('viewProject', $project);
+
         $okrsWithPage = $project->getOkrsWithPage($request);
+        $project['okrs'] = $okrsWithPage['okrs'];
 
         $data = [
-            'user' => auth()->user(),
-            'owner' => $project,
-            'okrs' => $okrsWithPage['okrs'],
+            'project' => $project,
             'pageInfo' => $okrsWithPage['pageInfo'],
             'st_date' => $request->input('st_date', ''),
             'fin_date' => $request->input('fin_date', ''),
@@ -152,7 +124,9 @@ class ProjectController extends Controller
 
     public function storeObjective(ObjectiveRequest $request, Project $project)
     {
-        $objective = $project->addObjective($request);
+        $this->authorize('storeObjective', $project);
+
+        $objective = $project->addObjective($request, $project);
         return redirect()->to(url()->previous() . '#oid-' . $objective->id);
     }
 
@@ -164,10 +138,13 @@ class ProjectController extends Controller
      */
     public function done(Project $project)
     {
+        $this->authorize('done', $project);
+
         $project->isdone = !$project->isdone;
         $project->save();
 
-        return redirect()->route('project');
+        if ($project->isdone) return redirect()->back();
+        else return redirect()->to(url()->previous() . '#closeProject');
     }
 
     /**
@@ -175,10 +152,35 @@ class ProjectController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function memberSetting(Project $project)
+    public function member(Request $request, Project $project)
     {
+        $builder = $project->users();
+
+        if ($request->input('order', '')) {
+            
+            # 排序
+            if ($order = $request->input('order', '')) { 
+                # 判斷value是以 _asc 或者 _desc 结尾來排序
+                if (preg_match('/^(.+)_(asc|desc)$/', $order, $m)) {
+                    # 判斷是否為指定的接收的參數
+                    if (in_array($m[1], ['name', 'email', 'position'])) {   
+                        # 開始排序              
+                        $builder->orderBy($m[1], $m[2]);
+                    }
+                }
+            }
+        } else {
+            # 預設
+            $builder->orderBy('id');
+        }
+
+        $pages = $builder->paginate(10)->appends([
+            'order' => $request->input('order', ''),
+        ]);
+
         $data = [
             'project' => $project,
+            'members' => $pages,
         ];
 
         return view('project.member', $data);
@@ -193,9 +195,11 @@ class ProjectController extends Controller
      */
     public function inviteMember(Request $request, Project $project)
     {
+        $this->authorize('memberSetting', $project);
+
         $project->sendInvitation($request);
 
-        return redirect()->route('project.member.setting', $project);
+        return redirect()->route('project.member', $project);
     }
 
     /**
@@ -207,9 +211,11 @@ class ProjectController extends Controller
      */
     public function cancelInvite(Project $project, User $member)
     {
+        $this->authorize('memberSetting', $project);
+
         $project->deleteInvitation($member);
 
-        return redirect()->route('project.member.setting', $project);
+        return redirect()->route('project.member', $project);
     }
 
     /**
@@ -236,7 +242,8 @@ class ProjectController extends Controller
     public function agreeInvite(Project $project, User $member)
     {
         $project->deleteInvitation($member);
-        ProjectUser::create(['project_id' => $project->id, 'user_id' => $member->id]);
+        $project->users()->attach($member);
+        $project->createPermission(3);
 
         return redirect()->route('project');
     }
@@ -262,6 +269,21 @@ class ProjectController extends Controller
     }
 
     /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function updateMember(Request $request, Project $project, User $member)
+    {
+        $this->authorize('memberSetting', $project);
+
+        if ($member->id != auth()->user()->id) $project->permissions()->where('user_id', $member->id)->update(['role_id' => $request->permission]);
+
+        return redirect()->route('project.member', $project);
+    }
+
+    /**
      * Remove department_id and position from storage.
      *
      * @param  \App\Project $project
@@ -270,8 +292,47 @@ class ProjectController extends Controller
      */
     public function destroyMember(Project $project, User $member)
     {
-        ProjectUser::where([['project_id', $project->id], ['user_id', $member->id]])->delete();
+        $this->authorize('memberSetting', $project);
+        Permission::where(['user_id' => $member->id, 'model_type' => Project::class, 'model_id' => $project->id])->delete();
+        $project->users()->detach($member);
 
-        return redirect()->route('project.member.setting', $project);
+        return redirect()->route('project.member', $project);
+    }
+
+    /**
+     * 變更最高權限管理者
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function changeAdmin(Request $request, Project $project)
+    {
+        $this->authorize('adminCange', [auth()->user(), $project]);
+
+        Permission::where(['user_id' => $request->invite, 'model_type' => Project::class, 'model_id' => $project->id])->update(['role_id' => 1]);
+        Permission::where(['user_id' => auth()->user()->id, 'model_type' => Project::class, 'model_id' => $project->id])->update(['role_id' => 2]);
+
+        return redirect()->back();
+    }
+
+    /**
+     * 刪除最高權限管理者
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function deleteAdmin(Request $request, Project $project)
+    {
+        $user = auth()->user();
+        $this->authorize('adminCange', [$user, $project]);
+
+        if ($request->project == $user->id) return redirect()->back();
+        else {
+            $permission->where('user_id', $request->project)->update(['role_id' => 1]);
+            $permission->model->users()->detach($user);
+            $permission->delete();
+        }
+
+        return redirect()->back();
     }
 }
